Refactor ColumnExplorer / ExplorerImpl
Extract an interface IQueryResolver. This way we can separate the
JsonApiClient and ExploreParams from the ExplorerImpl.
