+++ src/explorer.api/Explorers/AircloakQueryResolver.cs
@@ -1,0 +1,30 @@
+namespace Explorer
+{
+    using System;
+    using System.Threading.Tasks;
+
+    using Aircloak.JsonApi;
+    using Aircloak.JsonApi.ResponseTypes;
+    using Explorer.Api.Models;
+
+    internal class AircloakQueryResolver : IQueryResolver
+    {
+        public AircloakQueryResolver(JsonApiClient apiClient, string dataSourceName)
+        {
+            ApiClient = apiClient;
+            DataSourceName = dataSourceName;
+        }
+
+        public string DataSourceName { get; }
+
+        public JsonApiClient ApiClient { get; }
+
+        public async Task<QueryResult<TResult>> ResolveQuery<TResult>(IQuerySpec<TResult> query, TimeSpan timeout)
+        {
+            return await ApiClient.Query<TResult>(
+                DataSourceName,
+                query,
+                timeout);
+        }
+    }
+}
