+++ src/explorer.api/Explorers/IQueryResolver.cs
@@ -1,0 +1,14 @@
+namespace Explorer
+{
+    using System.Threading.Tasks;
+
+    using Aircloak.JsonApi;
+    using Aircloak.JsonApi.ResponseTypes;
+
+    internal interface IQueryResolver
+    {
+        public Task<QueryResult<TResult>> ResolveQuery<TResult>(
+            IQuerySpec<TResult> query,
+            System.TimeSpan timeout);
+    }
+}
